+++ app/controllers/application_controller.rb
@@ -3,6 +3,9 @@ class ApplicationController < ActionController::Base
   # Add theme folder to view path
   self.view_paths.unshift(*Rails.root.join('app', 'views', 'myapnea'))
 
+  # Add flash types
+  add_flash_types :warning
+
   # Layout
   layout 'layouts/application'