+++ app/models/user.rb
@@ -73,6 +73,7 @@ class User < ActiveRecord::Base
   has_many :posts, -> { where deleted: false }
   has_many :subscriptions
   has_many :users, class_name: "User", foreign_key: "provider_id"
+  has_many :invites
 
   # CDM
   has_one :cdm_demographic, foreign_key: 'patid'