+++ db/migrate/20150617150625_create_invites.rb
@@ -1,0 +1,12 @@
+class CreateInvites < ActiveRecord::Migration
+  def change
+    create_table :invites do |t|
+      t.string :email
+      t.integer :user_id
+      t.integer :recipient_id
+      t.string :token
+
+      t.timestamps null: false
+    end
+  end
+end